requirements.txt
wheel
requests
pyyaml
folium
tqdm
pyvis
networkx